requirements.txt
streamlit
pandas
requests
spotipy
beautifulsoup4
stqdm
python-dotenv
selenium
webdriver-manager
